+++ game/src/de/undefinedhuman/projectcreate/game/world/World.java
@@ -34,7 +34,7 @@ public class World {
 
     private Color batchColor = new Color();
 
-    private short[][][] blockLayer;       // Y, X, DATA LAYER (BLOCK, STATE)
+    private int[][][] blockLayer;       // Y, X, DATA LAYER (BLOCK, STATE)
     private short[][][] collisionLayer;   // Y, X, COLLISION LAYER
 
     public World(String name, int maxHeight, int width, int height, int seed) {
@@ -46,7 +46,7 @@ public class World {
         this.pixelSize.set(width * Variables.BLOCK_SIZE, height * Variables.BLOCK_SIZE);
         this.collisionSize.set(width*2, height*2);
 
-        this.blockLayer = new short[size.y][size.x][4];
+        this.blockLayer = new int[size.y][size.x][4];
         this.collisionLayer = new short[collisionSize.y][collisionSize.x][2];
     }
 
@@ -87,18 +87,18 @@ public class World {
             TopLayerManager.instance.render(batch, x, y, TopLayerType.GRASS, getBlock(x - 1, y, worldLayer) != 0, getBlock(x + 1, y, worldLayer) != 0);
     }
 
-    public short getBlock(int x, int y, byte worldLayer) {
+    public int getBlock(int x, int y, byte worldLayer) {
         if(y < 0 || y >= size.y) return 0;
         return getBlockData(x, y, worldLayer, BLOCK_LAYER);
     }
 
-   public void setBlock(int x, int y, byte worldLayer, short blockID) {
+   public void setBlock(int x, int y, byte worldLayer, int blockID) {
         if(y < 0 || y >= size.y) return;
         setBlockData(x, y, worldLayer, BLOCK_LAYER, blockID);
         if(worldLayer == MAIN_LAYER) setCollisionBlock(x, y, blockID != 0 && ((Block) ItemManager.getInstance().getItem(blockID)).hasCollision.getValue());
     }
 
-    public short getState(int x, int y, byte worldLayer) {
+    public int getState(int x, int y, byte worldLayer) {
         if(getBlock(x, y, worldLayer) == 0) return 0;
         return getBlockData(x, y, worldLayer, STATE_LAYER);
     }
@@ -124,11 +124,11 @@ public class World {
                 setCollision(blockX*2+i, blockY*2+j, COLLISION_BASE_LAYER, (byte) (solid ? 1 : 0));
     }
 
-    private short getBlockData(int x, int y, byte worldLayer, byte dataLayer) {
+    private int getBlockData(int x, int y, byte worldLayer, byte dataLayer) {
         return this.blockLayer[y][calculateXPosition(x, size.x)][worldLayer + dataLayer];
     }
 
-    private void setBlockData(int x, int y, byte worldLayer, byte dataLayer, short blockID) {
+    private void setBlockData(int x, int y, byte worldLayer, byte dataLayer, int blockID) {
         this.blockLayer[y][calculateXPosition(x, size.x)][worldLayer + dataLayer] = blockID;
     }
